src/configuration.py
from __future__ import annotations

import os
import re
from functools import lru_cache
from pathlib import Path
from typing import Any, Dict, Optional

import yaml
from dotenv import load_dotenv


def _project_root() -> Path:
    current = Path(__file__).resolve()
    for parent in [current] + list(current.parents):
        if (parent / "pyproject.toml").exists():
            return parent
    return current.parents[2]


def _substitute_env_vars(value: Any) -> Any:
    """Recursively substitute environment variables in config values.
    
    Syntax: ${VAR_NAME:default_value} or ${VAR_NAME}
    """
    if isinstance(value, str):
        pattern = r'\$\{([^}:]+)(?::([^}]*))?\}'
        def replacer(match: re.Match) -> str:
            var_name = match.group(1)
            default = match.group(2) or ""
            return os.getenv(var_name, default)
        return re.sub(pattern, replacer, value)
    elif isinstance(value, dict):
        return {k: _substitute_env_vars(v) for k, v in value.items()}
    elif isinstance(value, list):
        return [_substitute_env_vars(item) for item in value]
    return value


@lru_cache(maxsize=1)
def load_config(config_path: Optional[Path] = None) -> Dict[str, Any]:
    """Load and return the pipeline configuration."""
    load_dotenv()
    
    path = config_path or (_project_root() / "config" / "path.yaml")
    path = path.resolve()
    
    if not path.exists():
        raise FileNotFoundError(f"Pipeline config not found: {path}")
    
    with path.open("r", encoding="utf-8") as handle:
        config = yaml.safe_load(handle)
    
    return _substitute_env_vars(config)

def get_dbp_data_management_config(config: Optional[Dict[str, Any]] = None) -> Dict[str, str]:
    """Get dbp data management configuration."""

    cfg = config or load_config()
    data = {}
    data["raw"] = (_project_root() / cfg["data"] / "raw").resolve()
    data["cleaned"] = (_project_root() / cfg["data"] / "cleaned").resolve()

    return data

def get_suggestions_config(config: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
    """Get suggestions generation configuration with resolved paths."""
    cfg = config or load_config()
    case_name = get_case_name(cfg)
    suggestions = cfg["suggestions"].copy()
    
    llm_model_name = suggestions.get("llm_model", "unknown_model")
    llm_model_name_clean = re.sub(r'[^\w\-.]', '_', llm_model_name).strip('_')

    base_filename = suggestions.get("output_filename", "suggestions.json")
    name, ext = base_filename.rsplit(".", 1)
    output_filename = f"{name}_{llm_model_name_clean}.{ext}"
    
    suggestions["case_name"] = case_name
    suggestions["issues_path"] = (
        _project_root() / "data" / "processed" / "acc_result" / case_name / "issues" / "topics.json"
    ).resolve()
    suggestions["output_path"] = (
        _project_root() / "data" / "output" / case_name / output_filename
    ).resolve()
    
    return suggestions

