pyproject.toml
[project]
name = "dbp"
version = "0.1.0"
description = "digital building permit analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
